+++ stream-producer.py
@@ -7,14 +7,8 @@
 
 import argparse
 import asyncio
-import boto3
-#For testing
-from botocore import UNSIGNED
-from botocore.config import Config
 import collections
-import confluent_kafka
 import csv
-import fastavro
 import gzip
 import json
 import io
@@ -22,9 +16,6 @@ import linecache
 import logging
 import multiprocessing
 import os
-import pandas
-import pika
-import pyarrow.parquet as pq
 import queue
 import random
 import re
@@ -36,12 +27,19 @@ import threading
 import time
 import urllib.parse
 import urllib.request
-import websockets
+import boto3
+from botocore import UNSIGNED
+from botocore.config import Config
+import pika
+import confluent_kafka
+import fastavro
+import pandas
+import pyarrow.parquet as pq
 
 __all__ = []
-__version__ = "1.4.0"  # See https://www.python.org/dev/peps/pep-0396/
-__date__ = '2020-04-07'
-__updated__ = '2021-03-12'
+__version__ = "1.4.2"  # See https://www.python.org/dev/peps/pep-0396/
+__date__ = '2020-07-07'
+__updated__ = '2021-07-07'
 
 SENZING_PRODUCT_ID = "5014"  # See https://github.com/Senzing/knowledge-base/blob/master/lists/senzing-product-ids.md
 log_format = '%(asctime)s %(message)s'
@@ -165,6 +163,11 @@ configuration_locator = {
         "env": "SENZING_RABBITMQ_USERNAME",
         "cli": "rabbitmq-username",
     },
+    "rabbitmq_virtual_host": {
+        "default": pika.ConnectionParameters.DEFAULT_VIRTUAL_HOST,
+        "env": "SENZING_RABBITMQ_VIRTUAL_HOST",
+        "cli": "rabbitmq-virtual-host",
+    },
     "read_queue_maxsize": {
         "default": 50,
         "env": "SENZING_READ_QUEUE_MAXSIZE",
@@ -224,16 +227,6 @@ configuration_locator = {
         "env": "SENZING_THREADS_PER_PRINT",
         "cli": "threads-per-print"
     },
-    "websocket_host": {
-        "default": "0.0.0.0",
-        "env": "SENZING_WEBSOCKET_HOST",
-        "cli": "websocket-host"
-    },
-    "websocket_port": {
-        "default": 8255,
-        "env": "SENZING_WEBSOCKET_PORT",
-        "cli": "websocket-port"
-    }
 }
 
 # Enumerate keys in 'configuration_locator' that should not be printed to the log.
@@ -350,26 +343,6 @@ def get_parser():
         'parquet-to-stdout': {
             "help": 'Read Parquet file and print to STDOUT.',
             "argument_aspects": ["input-url", "parquet", "stdout"]
-        },
-        'websocket-to-kafka': {
-            "help": 'Read JSON from Websocket and send to Kafka.',
-            "argument_aspects": ["websocket", "kafka"]
-        },
-        'websocket-to-rabbitmq': {
-            "help": 'Read JSON from Websocket and send to RabbitMQ.',
-            "argument_aspects": ["websocket", "rabbitmq"]
-        },
-        'websocket-to-sqs': {
-            "help": 'Read JSON from Websocket and print to AWS SQS.',
-            "argument_aspects": ["websocket", "sqs"]
-        },
-        'websocket-to-sqs-batch': {
-            "help": 'Read JSON from Websocket and print to AWS SQS using batch.  DEPRECATED: Use websocket-to-sqs and set SENZING_RECORDS_PER_MESSAGE',
-            "argument_aspects": ["websocket", "sqs"]
-        },
-        'websocket-to-stdout': {
-            "help": 'Read JSON from Websocket and print to STDOUT.',
-            "argument_aspects": ["websocket", "stdout"]
         },
         'sleep': {
             "help": 'Do nothing but sleep. For Docker testing.',
@@ -497,24 +470,17 @@ def get_parser():
                 "metavar": "SENZING_RABBITMQ_USE_EXISTING_ENTITIES",
                 "help": "Connect to an existing exchange and queue using their settings. An error is thrown if the exchange or queue does not exist. If False, it will create the exchange and queue if they do not exist. If they exist, then it will attempt to connect, checking the settings match. Default: False"
             },
+            "--rabbitmq-virtual-host": {
+                "dest": "rabbitmq_virtual_host",
+                "metavar": "SENZING_RABBITMQ_VIRTUAL_HOST",
+                "help": "RabbitMQ virtual host. Default: None, which will use the RabbitMQ defined default virtual host"
+            },
         },
         "sqs": {
             "--sqs-queue-url": {
                 "dest": "sqs_queue_url",
                 "metavar": "SENZING_SQS_QUEUE_URL",
                 "help": "AWS SQS URL. Default: none"
-            },
-        },
-        "websocket": {
-            "--websocket-host": {
-                "dest": "websocket_host",
-                "metavar": "SENZING_WEBSOCKET_HOST",
-                "help": "Host to listen on. Default: 0.0.0.0"
-            },
-            "--websocket-port": {
-                "dest": "websocket_port",
-                "metavar": "SENZING_WEBSOCKET_PORT",
-                "help": "Port to listen on. Default: 8255"
             },
         },
         "csv": {
@@ -592,6 +558,7 @@ message_dictionary = {
     "299": "{0}",
     "300": "senzing-" + SENZING_PRODUCT_ID + "{0:04d}W",
     "310": "Did not send record identified by {0}: {1}. Exceeds SENZING_RECORD_SIZE_MAX by {2} bytes.",
+    "311": "Did not send record identified by {0}: {1}. Exceeds queue message size limit by {2} bytes.",
     "404": "Buffer error: {0} for line #{1} '{2}'.",
     "405": "Kafka error: {0} for line #{1} '{2}'.",
     "406": "Not implemented error: {0} for line #{1} '{2}'.",
@@ -643,7 +610,6 @@ def message(index, *args):
 
 
 def message_generic(generic_index, index, *args):
-    index_string = str(index)
     return "{0} {1}".format(message(generic_index, index), message(index, *args))
 
 
@@ -824,7 +790,7 @@ def redact_configuration(config):
     for key in keys_to_redact:
         try:
             result.pop(key)
-        except:
+        except Exception:
             pass
     return result
 
@@ -848,7 +814,7 @@ class Governor:
     def __enter__(self):
         return self
 
-    def __exit__(self, type, value, traceback):
+    def __exit__(self, exception_type, exception_value, exception_traceback):
         self.close()
 
 # -----------------------------------------------------------------------------
@@ -910,7 +876,8 @@ def exit_error(index, *args):
     ''' Log error message and exit program. '''
     logging.error(message_error(index, *args))
     logging.error(message_error(698))
-    sys.exit(1)
+    logging.shutdown()
+    os._exit(1)
 
 
 def exit_silently():
@@ -967,7 +934,7 @@ class MonitorThread(threading.Thread):
 
             interval_in_seconds = 5
             active_workers = len(self.workers)
-            for step in range(1, sleep_time_in_seconds, interval_in_seconds):
+            for _ in range(1, sleep_time_in_seconds, interval_in_seconds):
                 time.sleep(interval_in_seconds)
                 active_workers = len(self.workers)
                 for worker in self.workers:
@@ -1031,7 +998,7 @@ class MonitorThread(threading.Thread):
 
 class ReadFileAvroMixin():
 
-    def __init__(self, config={}, *args, **kwargs):
+    def __init__(self, config=None, *args, **kwargs):
         logging.debug(message_debug(996, threading.current_thread().name, "ReadFileAvroMixin"))
         self.input_url = config.get('input_url')
         self.record_min = config.get('record_min')
@@ -1056,7 +1023,7 @@ class ReadFileAvroMixin():
 
 class ReadFileCsvMixin():
 
-    def __init__(self, config={}, *args, **kwargs):
+    def __init__(self, config=None, *args, **kwargs):
         logging.debug(message_debug(996, threading.current_thread().name, "ReadFileCsvMixin"))
         self.input_url = config.get('input_url')
         self.record_min = config.get('record_min')
@@ -1078,7 +1045,7 @@ class ReadFileCsvMixin():
                     continue
                 if self.record_max and self.counter > self.record_max:
                     break
-                assert type(row) == dict
+                assert isinstance(row, dict)
                 yield row
 
 # -----------------------------------------------------------------------------
@@ -1088,7 +1055,7 @@ class ReadFileCsvMixin():
 
 class ReadFileGzippedMixin():
 
-    def __init__(self, config={}, *args, **kwargs):
+    def __init__(self, config=None, *args, **kwargs):
         logging.debug(message_debug(996, threading.current_thread().name, "ReadFileGzippedMixin"))
         self.input_url = config.get('input_url')
         self.record_min = config.get('record_min')
@@ -1116,8 +1083,8 @@ class ReadFileGzippedMixin():
 
 
 class ReadFileMixin():
-
-    def __init__(self, config={}, *args, **kwargs):
+  
+    def __init__(self, config=None, *args, **kwargs):
         logging.debug(message_debug(996, threading.current_thread().name, "ReadFileMixin"))
         self.input_url = config.get('input_url')
         self.record_min = config.get('record_min')
@@ -1137,7 +1104,7 @@ class ReadFileMixin():
                     continue
                 assert isinstance(line, str)
                 yield line
-                
+
 # -----------------------------------------------------------------------------
 # Class: ReadFileParquetMixin
 # -----------------------------------------------------------------------------
@@ -1145,7 +1112,7 @@ class ReadFileMixin():
 
 class ReadFileParquetMixin():
 
-    def __init__(self, config={}, *args, **kwargs):
+    def __init__(self, config=None, *args, **kwargs):
         logging.debug(message_debug(996, threading.current_thread().name, "ReadFileParquetMixin"))
         self.input_url = config.get('input_url')
         self.record_min = config.get('record_min')
@@ -1160,7 +1127,7 @@ class ReadFileParquetMixin():
                 continue
             if self.record_max and self.counter > self.record_max:
                 break
-            assert type(row) == dict
+            assert isinstance(row, dict)
             yield row
 
 # -----------------------------------------------------------------------------
@@ -1353,7 +1320,7 @@ class ReadS3ParquetMixin():
 
 class ReadUrlAvroMixin():
 
-    def __init__(self, config={}, *args, **kwargs):
+    def __init__(self, config=None, *args, **kwargs):
         logging.debug(message_debug(996, threading.current_thread().name, "ReadFileAvroMixin"))
         self.input_url = config.get('input_url')
         self.record_min = config.get('record_min')
@@ -1409,7 +1376,7 @@ class ReadUrlGzippedMixin():
 
 class ReadUrlMixin():
 
-    def __init__(self, config={}, *args, **kwargs):
+    def __init__(self, config=None, *args, **kwargs):
         logging.debug(message_debug(996, threading.current_thread().name, "ReadUrlMixin"))
         self.input_url = config.get('input_url')
         self.record_min = config.get('record_min')
@@ -1432,38 +1399,6 @@ class ReadUrlMixin():
             assert isinstance(result, dict)
             yield result
 
-# -----------------------------------------------------------------------------
-# Class: ReadWebsocketMixin
-# -----------------------------------------------------------------------------
-
-
-class ReadWebsocketMixin():
-
-    def __init__(self, config={}, *args, **kwargs):
-        logging.debug(message_debug(996, threading.current_thread().name, "ReadWebsocketMixin"))
-        self.websocket_host = config.get("websocket_host")
-        self.websocket_port = config.get("websocket_port")
-        self.local_queue = multiprocessing.Queue()
-
-        # Start websocket server.
-
-        self.start_server = websockets.serve(self.websocket_server_handler, self.websocket_host, self.websocket_port)
-        asyncio.get_event_loop().run_until_complete(self.start_server)
-#         asyncio.get_event_loop().run_forever()
-
-    async def websocket_server_handler(self, websocket, path):
-        async for record in websocket:
-            self.local_queue.put(record)
-
-    def read(self):
-        while True:
-            record = self.local_queue.get(block=True)
-            yield record
-
-        # Cleanup, if "while True" ever changes.
-
-        self.local_queue.close()
-        self.local_queue.join_thread()
 
 # =============================================================================
 # Mixins: Evaluate*
@@ -1540,7 +1475,7 @@ class EvaluateMakeSerializeableDictMixin():
             try:
                 if value.isnumeric():
                     new_message[key] = value
-            except:
+            except Exception:
                 pass
         return new_message
 
@@ -1561,7 +1496,7 @@ class EvaluateMakeSerializeableDictMixin():
 
 class PrintKafkaMixin():
 
-    def __init__(self, config={}, *args, **kwargs):
+    def __init__(self, config=None, *args, **kwargs):
         logging.debug(message_debug(996, threading.current_thread().name, "PrintKafkaMixin"))
         self.config = config
         self.kafka_poll_interval = config.get("kafka_poll_interval")
@@ -1570,6 +1505,14 @@ class PrintKafkaMixin():
         self.number_of_records_per_print = config.get("records_per_message")
         self.message_buffer = '['
         self.num_messages = 0
+
+        # default Kafka max message size, but save some space for kafka to use, just in case
+
+        self.max_message_size_in_bytes = 1000012 - (32 * 1024)
+
+
+        if self.number_of_records_per_print <= 0:
+            self.number_of_records_per_print = sys.maxsize
 
         kafka_configuration = {
             'bootstrap.servers': config.get('kafka_bootstrap_server')
@@ -1587,6 +1530,21 @@ class PrintKafkaMixin():
     def print(self, message):
         assert isinstance(message, str)
 
+        # if the record itself is too long to fit in a message, then log the ID with a warning and move on (+2 for the enclosing [])
+
+        new_record_size_in_bytes = len(message.encode('utf-8'))
+        if new_record_size_in_bytes + 2 > self.max_message_size_in_bytes:
+            record = json.dumps(message)
+            record_id = record.get(self.record_identifier)
+            record_overage = new_record_size_in_bytes + 2 - self.max_message_size_in_bytes
+            logging.warning(message_warning(311, self.record_identifier, record_id, record_overage))
+            return
+
+        # Check if the new record would overflow the message and if so, send the existing messages
+
+        if len(self.message_buffer.encode('utf-8')) + new_record_size_in_bytes + 1 > self.max_message_size_in_bytes:
+            self.send_message_buffer()
+
         # batch the message - if are already messages then add a delimiter first
 
         if self.num_messages > 0:
@@ -1596,22 +1554,15 @@ class PrintKafkaMixin():
 
         try:
             if self.num_messages == self.number_of_records_per_print:
-                self.message_buffer += ']'
-                self.kafka_producer.produce(
-                    self.kafka_topic,
-                    self.message_buffer,
-                    on_delivery=self.on_kafka_delivery
-                )
-                self.message_buffer = '['
-                self.num_messages = 0
+                self.send_message_buffer()
 
         except BufferError as err:
             logging.warning(message_warning(404, err, message))
         except confluent_kafka.KafkaException as err:
             logging.warning(message_warning(405, err, message))
-        except NotImplemented as err:
+        except NotImplementedError as err:
             logging.warning(message_warning(406, err, message))
-        except:
+        except Exception:
             logging.warning(message_warning(407, err, message))
 
         # Log progress. Using a "cheap" serialization technique.
@@ -1629,15 +1580,18 @@ class PrintKafkaMixin():
 
     def close(self):
         if self.num_messages > 0:
-            self.message_buffer += ']'
-            self.kafka_producer.produce(
+            self.send_message_buffer()
+        self.kafka_producer.flush()
+
+    def send_message_buffer(self):
+        self.message_buffer += ']'
+        self.kafka_producer.produce(
                     self.kafka_topic,
                     self.message_buffer,
                     on_delivery=self.on_kafka_delivery
                 )
-            self.message_buffer = ''
-            self.num_messages = 0
-        self.kafka_producer.flush()
+        self.message_buffer = '['
+        self.num_messages = 0
 
 # -----------------------------------------------------------------------------
 # Class: PrintRabbitmqMixin
@@ -1646,7 +1600,7 @@ class PrintKafkaMixin():
 
 class PrintRabbitmqMixin():
 
-    def __init__(self, config={}, *args, **kwargs):
+    def __init__(self, config=None, *args, **kwargs):
         logging.debug(message_debug(996, threading.current_thread().name, "PrintRabbitmqMixin"))
 
         rabbitmq_delivery_mode = 2
@@ -1655,6 +1609,7 @@ class PrintRabbitmqMixin():
         rabbitmq_port = config.get("rabbitmq_port")
         rabbitmq_username = config.get("rabbitmq_username")
         rabbitmq_passive_declare = config.get("rabbitmq_use_existing_entities")
+        rabbitmq_virtual_host = config.get("rabbitmq_virtual_host")
         self.rabbitmq_exchange = config.get("rabbitmq_exchange")
         self.rabbitmq_queue = config.get("rabbitmq_queue")
         self.rabbitmq_routing_key = config.get("rabbitmq_routing_key")
@@ -1662,6 +1617,13 @@ class PrintRabbitmqMixin():
         self.number_of_records_per_print = config.get("records_per_message")
         self.message_buffer = '['
         self.num_messages = 0
+
+        # default RabbitMQ max message size, but save some space for Rabbit to use, just in case
+
+        self.max_message_size_in_bytes = (128 * 1024 * 1024) - (32 * 1024)
+
+        if self.number_of_records_per_print <= 0:
+            self.number_of_records_per_print = sys.maxsize
 
         # Construct Pika objects.
 
@@ -1675,7 +1637,8 @@ class PrintRabbitmqMixin():
         rabbitmq_connection_parameters = pika.ConnectionParameters(
             host=rabbitmq_host,
             port=rabbitmq_port,
-            credentials=credentials
+            credentials=credentials,
+            virtual_host=rabbitmq_virtual_host
         )
 
         # Open connection to RabbitMQ.
@@ -1687,6 +1650,7 @@ class PrintRabbitmqMixin():
             message_queue = self.channel.queue_declare(queue=self.rabbitmq_queue, passive=rabbitmq_passive_declare)
 
             # if we are actively declaring, then we need to bind. If passive declare, we assume it is already set up
+
             if not rabbitmq_passive_declare:
                 self.channel.queue_bind(exchange=self.rabbitmq_exchange, routing_key=self.rabbitmq_routing_key, queue=message_queue.method.queue)
         except (pika.exceptions.AMQPConnectionError) as err:
@@ -1704,7 +1668,23 @@ class PrintRabbitmqMixin():
     def print(self, message):
         assert isinstance(message, str)
 
+        # if the record itself is too long to fit in a message, then log the ID with a warning and move on (+2 for the enclosing [])
+
+        new_record_size_in_bytes = len(message.encode('utf-8'))
+        if new_record_size_in_bytes + 2 > self.max_message_size_in_bytes:
+            record = json.dumps(message)
+            record_id = record.get(self.record_identifier)
+            record_overage = new_record_size_in_bytes + 2 - self.max_message_size_in_bytes
+            logging.warning(message_warning(311, self.record_identifier, record_id, record_overage))
+            return
+
+        # Check if the new record would overflow the message and if so, send the existing messages
+
+        if len(self.message_buffer.encode('utf-8')) + new_record_size_in_bytes + 1 > self.max_message_size_in_bytes:
+            self.send_message_buffer()
+
         # batch the message - if are already messages then add a delimiter first
+
         if self.num_messages > 0:
             self.message_buffer += ','
         self.message_buffer += message
@@ -1714,18 +1694,10 @@ class PrintRabbitmqMixin():
 
         try:
             if self.num_messages == self.number_of_records_per_print:
-                self.message_buffer += ']'
-                self.channel.basic_publish(
-                    exchange=self.rabbitmq_exchange,
-                    routing_key=self.rabbitmq_routing_key,
-                    body=self.message_buffer,
-                    properties=self.rabbitmq_properties
-                )
-                self.message_buffer = '['
-                self.num_messages = 0
+                self.send_message_buffer()
 
         except BaseException as err:
-            logging.warn(message_warning(411, err, message))
+            logging.warning(message_warning(411, err, message))
 
         # Log progress. Using a "cheap" serialization technique.
         output_counter = self.config.get('output_counter')
@@ -1736,17 +1708,20 @@ class PrintRabbitmqMixin():
 
     def close(self):
         if self.num_messages > 0:
-            self.message_buffer += ']'
-            self.channel.basic_publish(
-                exchange=self.rabbitmq_exchange,
-                routing_key=self.rabbitmq_routing_key,
-                body=self.message_buffer,
-                properties=self.rabbitmq_properties
-            )
-            self.message_buffer = ''
-            self.num_messages = 0
+            self.send_message_buffer()
 
         self.connection.close()
+
+    def send_message_buffer(self):
+        self.message_buffer += ']'
+        self.channel.basic_publish(
+            exchange=self.rabbitmq_exchange,
+            routing_key=self.rabbitmq_routing_key,
+            body=self.message_buffer,
+            properties=self.rabbitmq_properties
+        )
+        self.message_buffer = '['
+        self.num_messages = 0
 
 # -----------------------------------------------------------------------------
 # Class: PrintQueueMixin
@@ -1783,6 +1758,10 @@ class PrintSqsMixin():
         self.number_of_records_per_print = config.get("records_per_message")
         self.message_buffer = '['
         self.num_messages = 0
+        self.record_identifier = config.get("record_identifier")
+
+        if self.number_of_records_per_print <= 0:
+            self.number_of_records_per_print = sys.maxsize
 
         # Create sqs object.
         # https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/sqs.html
@@ -1796,41 +1775,59 @@ class PrintSqsMixin():
         endpoint_url = match.group(1)
         self.sqs = boto3.client("sqs", endpoint_url=endpoint_url)
 
+        # Query queue for max message size, then leave some overhead space in case SQS needs some itself
+
+        response = self.sqs.get_queue_attributes(QueueUrl=self.queue_url, AttributeNames=['MaximumMessageSize'])
+        self.max_message_size_in_bytes = response.get('Attributes', { "MaximumMessageSize" : 256 * 1024}).get('MaximumMessageSize')
+        self.max_message_size_in_bytes = int(self.max_message_size_in_bytes) - (32 * 1024)
+
     def print(self, message):
         self.counter += 1
-        assert type(message) == str
+        assert isinstance(message, str)
+
+        # if the record itself is too long to fit in a message, then log the ID with a warning and move on (+2 for the enclosing [])
+
+        new_record_size_in_bytes = len(message.encode('utf-8'))
+        if new_record_size_in_bytes + 2 > self.max_message_size_in_bytes:
+            record = json.dumps(message)
+            record_id = record.get(self.record_identifier)
+            record_overage = new_record_size_in_bytes + 2 - self.max_message_size_in_bytes
+            logging.warning(message_warning(311, self.record_identifier, record_id, record_overage))
+            return
+
+        # Check if the new record would overflow the message and if so, send the existing messages
+
+        if len(self.message_buffer.encode('utf-8')) + new_record_size_in_bytes + 1 > self.max_message_size_in_bytes:
+            self.send_message_buffer()
 
         # batch the message - if are already messages then add a delimiter first
+
         if self.num_messages > 0:
             self.message_buffer += ','
         self.message_buffer += message
         self.num_messages += 1
 
         if self.num_messages == self.number_of_records_per_print:
-            self.message_buffer += ']'
-            response = self.sqs.send_message(
-                QueueUrl=self.queue_url,
-                DelaySeconds=self.sqs_delay_seconds,
-                MessageAttributes={},
-                MessageBody=(self.message_buffer),
-            )
-            self.message_buffer = '['
-            self.num_messages = 0
+            self.send_message_buffer()
 
         if self.counter % self.record_monitor == 0:
             logging.info(message_debug(104, threading.current_thread().name, self.counter))
 
     def close(self):
         if self.num_messages > 0:
-            self.message_buffer += ']'
-            response = self.sqs.send_message(
-                QueueUrl=self.queue_url,
-                DelaySeconds=self.sqs_delay_seconds,
-                MessageAttributes={},
-                MessageBody=(self.message_buffer),
-            )
-            self.message_buffer = ''
-            self.num_messages = 0
+            self.send_message_buffer()
+
+    def send_message_buffer(self):
+        self.message_buffer += ']'
+        self.sqs.send_message(
+            QueueUrl=self.queue_url,
+            DelaySeconds=self.sqs_delay_seconds,
+            MessageAttributes={},
+            MessageBody=(self.message_buffer),
+        )
+        self.message_buffer = '['
+        self.num_messages = 0
+
 
 # -----------------------------------------------------------------------------
 # Class: PrintSqsBatchMixin
@@ -1862,7 +1859,7 @@ class PrintSqsBatchMixin():
 
     def print(self, message):
         self.counter += 1
-        assert type(message) == str
+        assert isinstance(message, str)
         self.messages.append(message)
         if len(self.messages) >= 10:
             entries = []
@@ -1873,7 +1870,7 @@ class PrintSqsBatchMixin():
                     "DelaySeconds": self.sqs_delay_seconds
                 }
                 entries.append(entry)
-            response = self.sqs.send_message_batch(
+            self.sqs.send_message_batch(
                 QueueUrl=self.queue_url,
                 Entries=entries,
             )
@@ -1891,7 +1888,7 @@ class PrintSqsBatchMixin():
             }
             entries.append(entry)
         if len(entries) > 0:
-            response = self.sqs.send_message_batch(
+            self.sqs.send_message_batch(
                 QueueUrl=self.queue_url,
                 Entries=entries,
             )
@@ -1912,7 +1909,7 @@ class PrintStdoutMixin():
 
     def print(self, message):
         self.counter += 1
-        assert type(message) == str
+        assert isinstance(message, str)
         print(message)
         if self.counter % self.record_monitor == 0:
             logging.info(message_debug(104, threading.current_thread().name, self.counter))
@@ -1948,7 +1945,7 @@ class ReadEvaluatePrintLoopThread(threading.Thread):
         return self.governor.govern()
 
     def log_excessive_record(self, record, record_json):
-        assert type(record_json) == str
+        assert isinstance(record_json, str)
         record_overage = len(record_json) - self.record_size_max
         record_id = record.get(self.record_identifier)
         logging.warning(message_warning(310, self.record_identifier, record_id, record_overage))
@@ -2131,123 +2128,18 @@ class FilterUrlJsonToDictQueueThread(ReadEvaluatePrintLoopThread, ReadUrlMixin, 
             base.__init__(self, *args, **kwargs)
 
 
-class FilterWebsocketToDictQueueThread(ReadEvaluatePrintLoopThread, ReadWebsocketMixin, EvaluateJsonToDictMixin, PrintQueueMixin):
-
-    def __init__(self, *args, **kwargs):
-        logging.debug(message_debug(997, threading.current_thread().name, "FilterUrlJsonToDictQueueThread"))
-        for base in type(self).__bases__:
-            base.__init__(self, *args, **kwargs)
-
 # -----------------------------------------------------------------------------
 # *_processor
 # -----------------------------------------------------------------------------
 
 
-def pipeline_runner(
-    args=None,
-    options_to_defaults_map={},
-    pipeline=[],
-    monitor_thread=None,
-):
-
-    # Get context from CLI, environment variables, and ini files.
-
-    config = get_configuration(args)
-    validate_configuration(config)
-
-    # If configuration values not specified, use defaults.
-
-    for key, value in options_to_defaults_map.items():
-        if not config.get(key):
-            config[key] = config.get(value)
-
-    # Prolog.
-
-    logging.info(entry_template(config))
-
-    # If requested, delay start.
-
-    delay(config)
-
-    # Pull values from configuration.
-
-    default_queue_maxsize = config.get('read_queue_maxsize')
-
-    # Create threads for master process.
-
-    threads = []
-    input_queue = None
-
-    # Create pipeline segments.
-
-    for filter in pipeline:
-
-        # Get metadata about the filter.
-
-        filter_class = filter.get("class")
-        filter_threads = filter.get("threads", 1)
-        filter_queue_max_size = filter.get("queue_max_size", default_queue_maxsize)
-        filter_counter_name = filter.get("counter_name")
-        filter_delay = filter.get("delay", 1)
-
-        # Give prior filter a head start
-
-        time.sleep(filter_delay)
-
-        # Create internal Queue.
-
-        output_queue = multiprocessing.Queue(filter_queue_max_size)
-
-        # Start threads.
-
-        for i in range(0, filter_threads):
-            thread = filter_class(
-                config=config,
-                counter_name=filter_counter_name,
-                input_queue=input_queue,
-                output_queue=output_queue,
-            )
-            thread.name = "Process-0-{0}-{1}".format(thread.__class__.__name__, i)
-            threads.append(thread)
-            thread.start()
-
-        # Prepare for next filter.
-
-        input_queue = output_queue
-
-    # Add a monitoring thread.
-
-    adminThreads = []
-
-    if monitor_thread:
-        thread = monitor_thread(
-            config=config,
-            workers=threads,
-        )
-        thread.name = "Process-0-{0}-0".format(thread.__class__.__name__)
-        adminThreads.append(thread)
-        thread.start()
-
-    # Collect inactive threads.
-
-    for thread in threads:
-        thread.join()
-    for thread in adminThreads:
-        thread.join()
-
-    # Epilog.
-
-    logging.info(exit_template(config))
-
-
 def pipeline_read_write(
     args=None,
-    options_to_defaults_map={},
+    options_to_defaults_map=None,
     read_thread=None,
     write_thread=None,
     monitor_thread=None,
     governor=None,
-    run_async=False,
 ):
 
     # Get context from CLI, environment variables, and ini files.
@@ -2257,9 +2149,10 @@ def pipeline_read_write(
 
     # If configuration values not specified, use defaults.
 
-    for key, value in options_to_defaults_map.items():
-        if not config.get(key):
-            config[key] = config.get(value)
+    if options_to_defaults_map is not None:
+        for key, value in options_to_defaults_map.items():
+            if not config.get(key):
+                config[key] = config.get(value)
 
     # Prolog.
 
@@ -2326,11 +2219,6 @@ def pipeline_read_write(
         adminThreads.append(thread)
         thread.start()
 
-    # WebSocket requires an asyncio loop.
-
-    if run_async:
-        asyncio.get_event_loop().run_forever()
-
     # Collect inactive threads.
 
     for thread in threads:
@@ -2387,12 +2275,6 @@ def dohelper_avro(args, write_thread):
 
 def dohelper_csv(args, write_thread):
     ''' Read file of CSV, print to write_thread. '''
-
-    # Get context variables.
-
-    config = get_configuration(args)
-    input_url = config.get("input_url")
-    parsed_file_name = urllib.parse.urlparse(input_url)
 
     # Determine Read thread.
 
@@ -2495,12 +2377,6 @@ def dohelper_json(args, write_thread):
 def dohelper_parquet(args, write_thread):
     ''' Read file of Parquet, print to write_thread. '''
 
-    # Get context variables.
-
-    config = get_configuration(args)
-    input_url = config.get("input_url")
-    parsed_file_name = urllib.parse.urlparse(input_url)
-
     # Determine Read thread.
 
     read_thread = FilterFileParquetToDictQueueThread
@@ -2526,37 +2402,6 @@ def dohelper_parquet(args, write_thread):
         governor=governor
     )
 
-
-def dohelper_websocket(args, write_thread):
-    ''' Read file of JSON, print to write_thread. '''
-
-    # Get context variables.
-
-    config = get_configuration(args)
-
-    # Determine Read thread.
-
-    read_thread = FilterWebsocketToDictQueueThread  # Default.
-
-    # Cascading defaults.
-
-    options_to_defaults_map = {}
-
-    # Create governor.
-
-    governor = Governor(hint="stream-producer")
-
-    # Run pipeline.
-
-    pipeline_read_write(
-        args=args,
-        options_to_defaults_map=options_to_defaults_map,
-        read_thread=read_thread,
-        write_thread=write_thread,
-        monitor_thread=MonitorThread,
-        governor=governor,
-        run_async=True
-    )
 
 # -----------------------------------------------------------------------------
 # do_* functions
@@ -2768,35 +2613,6 @@ def do_version(args):
     logging.info(message_info(294, __version__, __updated__))
 
 
-def do_websocket_to_kafka(args):
-    ''' Read JSON from Websocket, print to Kafka. '''
-    write_thread = FilterQueueDictToJsonKafkaThread
-    dohelper_websocket(args, write_thread)
-
-
-def do_websocket_to_rabbitmq(args):
-    ''' Read JSON from Websocket, print to RabbitMQ. '''
-    write_thread = FilterQueueDictToJsonRabbitmqThread
-    dohelper_websocket(args, write_thread)
-
-
-def do_websocket_to_sqs(args):
-    ''' Read JSON from Websocket, print to AWS SQS. '''
-    write_thread = FilterQueueDictToJsonSqsThread
-    dohelper_websocket(args, write_thread)
-
-
-def do_websocket_to_sqs_batch(args):
-    ''' Read JSON from Websocket, print to AWS SQS. '''
-    write_thread = FilterQueueDictToJsonSqsBatchThread
-    dohelper_websocket(args, write_thread)
-
-
-def do_websocket_to_stdout(args):
-    ''' Read JSON from Websocket, print to STDOUT. '''
-    write_thread = FilterQueueDictToJsonStdoutThread
-    dohelper_websocket(args, write_thread)
-
 # -----------------------------------------------------------------------------
 # Main
 # -----------------------------------------------------------------------------
@@ -2847,8 +2663,7 @@ if __name__ == "__main__":
     else:
         parser.print_help()
         if len(os.getenv("SENZING_DOCKER_LAUNCHED", "")):
-            subcommand = "sleep"
-            args = argparse.Namespace(subcommand=subcommand)
+            args = argparse.Namespace(subcommand='sleep')
             do_sleep(args)
         exit_silently()